Added highlighting for points
The scrap highlight is quickly made to work.  There is a way to get it
to work putting the scrap img in the CSS and using linear-gradient for a
color overlay.  Not sure if we want that thinking big picture because as
we re-design the front end, notification of where you play a card might
change.